challange2/requirements.txt
google-cloud-bigquery
flask
gunicorn